Conflicts:
	pandora_console/extras/mr/46.sql

+++ pandora_console/godmode/alerts/alert_commands.php
@@ -28,6 +28,12 @@ if (! check_acl($config['id_user'], 0, 'LM')) {
     exit;
 }
 
+if (!check_acl($config['id_user'], 0, 'PM') && !is_user_admin($config['id_user'])) {
+    echo "<div id='message_permissions'  title='".__('Permissions warning')."' style='display:none;'>";
+    echo "<p style='text-align: center;font-weight: bold;'>".__('Command management is limited to administrator users or user profiles with permissions over Pandora FMS management').'</p>';
+    echo '</div>';
+}
+
 if (is_metaconsole()) {
     $sec = 'advanced';
 } else {
@@ -39,6 +45,8 @@ $update_command = (bool) get_parameter('update_command');
 $create_command = (bool) get_parameter('create_command');
 $delete_command = (bool) get_parameter('delete_command');
 $copy_command = (bool) get_parameter('copy_command');
+
+$url = 'index.php?sec='.$sec.'&sec2=godmode/alerts/alert_commands';
 
 if (is_ajax()) {
     $get_alert_command = (bool) get_parameter('get_alert_command');
@@ -56,7 +64,7 @@ if (is_ajax()) {
 
         $command = alerts_get_alert_command($id);
 
-        // If is setted a description, we change the carriage return by <br> tags
+        // If a description is set, change the carriage return by <br> tags.
         if (isset($command['description'])) {
             $command['description'] = str_replace(
                 [
@@ -90,9 +98,15 @@ if (is_ajax()) {
             if (!empty($field_description)) {
                 // If the value is 5,  this because severity in snmp alerts is not permit to show.
                 if (($i > 5) && ($command['id'] == 3)) {
-                    $fdesc = $field_description.' <br><span style="font-size:xx-small; font-weight:normal;">'.sprintf(__('Field %s'), ($i - 1)).'</span>';
+                    $fdesc = $field_description.' <br><span class="normal xx-small">'.sprintf(
+                        __('Field %s'),
+                        ($i - 1)
+                    ).'</span>';
                 } else {
-                    $fdesc = $field_description.' <br><span style="font-size:xx-small; font-weight:normal;">'.sprintf(__('Field %s'), $i).'</span>';
+                    $fdesc = $field_description.' <br><span class="normal xx-small">'.sprintf(
+                        __('Field %s'),
+                        $i
+                    ).'</span>';
                 }
 
                 // If the field is the number one, print the help message.
@@ -125,7 +139,7 @@ if (is_ajax()) {
                 $field_value = io_safe_output($field_value);
                 // HTML type.
                 if (preg_match('/^_html_editor_$/i', $field_value)) {
-                    $editor_type_chkbx = '<div style="padding: 4px 0px;"><b><small>';
+                    $editor_type_chkbx = '<div id="command_div"><b><small>';
                     $editor_type_chkbx .= __('Basic');
                     $editor_type_chkbx .= ui_print_help_tip(
                         __('For sending emails, text must be HTML format, if you want to use plain text, type it between the following labels: <pre></pre>'),
@@ -166,7 +180,7 @@ if (is_ajax()) {
                         $is_central_policies_on_node
                     );
 
-                    $editor_type_chkbx = '<div style="padding: 4px 0px;"><b><small>';
+                    $editor_type_chkbx = '<div id="command_div"><b><small>';
                     $editor_type_chkbx .= __('Basic').'&nbsp;&nbsp;';
                     $editor_type_chkbx .= html_print_radio_button_extended(
                         'editor_type_recovery_value_'.$i,
@@ -203,7 +217,7 @@ if (is_ajax()) {
                         $is_central_policies_on_node
                     );
                 } else if (preg_match('/^_content_type_$/i', $field_value)) {
-                    $editor_type_chkbx = '<div style="padding: 4px 0px;"><b><small>';
+                    $editor_type_chkbx = '<div id="command_div"><b><small>';
                     $editor_type_chkbx .= __('Text/plain');
                     $editor_type_chkbx .= ui_print_help_tip(
                         __('For sending emails only text plain'),
@@ -234,7 +248,7 @@ if (is_ajax()) {
                     $editor_type_chkbx .= '</small></b></div>';
                     $ffield = $editor_type_chkbx;
 
-                    $editor_type_chkbx = '<div style="padding: 4px 0px;"><b><small>';
+                    $editor_type_chkbx = '<div id="command_div"><b><small>';
                     $editor_type_chkbx .= __('Text/plain');
                     $editor_type_chkbx .= ui_print_help_tip(
                         __('For sending emails only text plain'),
@@ -265,6 +279,76 @@ if (is_ajax()) {
                     $editor_type_chkbx .= '</small></b></div>';
                     $rfield = $editor_type_chkbx;
                     // Select type.
+                } else if (preg_match('/^_integria_type_custom_field_$/i', $field_value)) {
+                        $ffield = '';
+                        $rfield = '';
+
+                        $ffield .= '<div name="field'.$i.'_value_container">'.html_print_switch(
+                            [
+                                'name'  => 'field'.$i.'_value[]',
+                                'value' => '',
+                            ]
+                        ).'</div>';
+                        $rfield .= '<div name="field'.$i.'_recovery_value_container">'.html_print_switch(
+                            [
+                                'name'  => 'field'.$i.'_recovery_value[]',
+                                'value' => '',
+                            ]
+                        ).'</div>';
+
+                        $ffield .= html_print_select(
+                            '',
+                            'field'.$i.'_value[]',
+                            '',
+                            '',
+                            __('None'),
+                            '',
+                            true,
+                            false,
+                            false,
+                            'fields',
+                            $is_central_policies_on_node
+                        );
+
+                        $rfield .= html_print_select(
+                            '',
+                            'field'.$i.'_recovery_value[]',
+                            '',
+                            '',
+                            __('None'),
+                            '',
+                            true,
+                            false,
+                            false,
+                            'fields',
+                            $is_central_policies_on_node
+                        );
+
+                        $ffield .= html_print_input_text('field'.$i.'_value[]', '', '', 10, 10, true, false, false, '', 'datepicker');
+                        $rfield .= html_print_input_text('field'.$i.'_recovery_value[]', '', '', 10, 10, true, false, false, '', 'datepicker');
+
+                        $ffield .= html_print_textarea(
+                            'field'.$i.'_value[]',
+                            1,
+                            1,
+                            '',
+                            'style="min-height:40px; '.$style.'" class="fields"',
+                            true,
+                            '',
+                            $is_central_policies_on_node
+                        );
+
+
+                        $rfield .= html_print_textarea(
+                            'field'.$i.'_recovery_value[]',
+                            1,
+                            1,
+                            '',
+                            'style="min-height:40px; '.$style.'" class="fields_recovery',
+                            true,
+                            '',
+                            $is_central_policies_on_node
+                        );
                 } else {
                     $fields_value_select = [];
                     $fv = explode(';', $field_value);
@@ -304,7 +388,7 @@ if (is_ajax()) {
                             'field'.$i.'_recovery_value',
                             '',
                             '',
-                            '',
+                            __('None'),
                             0,
                             true,
                             false,
@@ -318,7 +402,7 @@ if (is_ajax()) {
                             1,
                             1,
                             $fv[0],
-                            'style="min-height:40px; '.$style.'" class="fields"',
+                            'style="'.$style.'" class="fields min-height-40px"',
                             true,
                             '',
                             $is_central_policies_on_node
@@ -328,7 +412,7 @@ if (is_ajax()) {
                             1,
                             1,
                             $fv[0],
-                            'style="min-height:40px; '.$style.'" class="fields_recovery',
+                            'style="'.$style.'" class="fields_recovery min-height-40px',
                             true,
                             '',
                             $is_central_policies_on_node
@@ -341,7 +425,7 @@ if (is_ajax()) {
                     1,
                     1,
                     '',
-                    'style="min-height:40px; '.$style.'" class="fields"',
+                    'style="'.$style.'" class="fields min-height-40px"',
                     true,
                     '',
                     $is_central_policies_on_node
@@ -351,7 +435,7 @@ if (is_ajax()) {
                     1,
                     1,
                     '',
-                    'style="min-height:40px; '.$style.'" class="fields_recovery"',
+                    'style="'.$style.'" class="fields_recovery min-height-40px"',
                     true,
                     '',
                     $is_central_policies_on_node
@@ -364,7 +448,7 @@ if (is_ajax()) {
                 $fields_rows[$i] = '';
             } else {
                 $fields_rows[$i] = '<tr id="table_macros-field'.$i.'" class="datos">';
-                $fields_rows[$i] .= '<td style="font-weight:bold;width:20%" class="datos">'.$fdesc.'</td>';
+                $fields_rows[$i] .= '<td class="datos bolder w20p">'.$fdesc.'</td>';
                 $fields_rows[$i] .= '<td class="datos">'.$ffield.'</td>';
                 if ($get_recovery_fields) {
                     $fields_rows[$i] .= '<td class="datos recovery_col">'.$rfield.'</td>';
@@ -552,11 +636,19 @@ if ($commands === false) {
     $commands = [];
 }
 
+// Pagination.
+$total_commands = count($commands);
+$offset = (int) get_parameter('offset');
+$limit = (int) $config['block_size'];
+$commands = array_slice($commands, $offset, $limit);
+
 foreach ($commands as $command) {
     $data = [];
 
-    $data['name'] = '<span style="font-size: 7.5pt">';
-    if (! $command['internal']) {
+    $data['name'] = '<span>';
+
+    // (IMPORTANT, DO NOT CHANGE!) only users with permissions over "All" group have access to edition of commands belonging to "All" group.
+    if (!$command['internal'] && check_acl_restricted_all($config['id_user'], $command['id_group'], 'PM')) {
         $data['name'] .= '<a href="index.php?sec='.$sec.'&sec2=godmode/alerts/configure_alert_command&id='.$command['id'].'&pure='.$pure.'">'.$command['name'].'</a>';
     } else {
         $data['name'] .= $command['name'];
@@ -580,25 +672,34 @@ foreach ($commands as $command) {
     );
     $data['action'] = '';
     $table->cellclass[]['action'] = 'action_buttons';
-    if ($is_central_policies_on_node === false && !$command['internal']) {
-        $data['action'] = '<span style="display: inline-flex">';
+
+    // (IMPORTANT, DO NOT CHANGE!) only users with permissions over "All" group have access to edition of commands belonging to "All" group.
+    if ($is_central_policies_on_node === false && !$command['internal'] && check_acl_restricted_all($config['id_user'], $command['id_group'], 'LM')) {
+        $data['action'] = '<span class="inline_flex">';
         $data['action'] .= '<a href="index.php?sec='.$sec.'&sec2=godmode/alerts/alert_commands&amp;copy_command=1&id='.$command['id'].'&pure='.$pure.'"
-			onClick="if (!confirm(\''.__('Are you sure?').'\')) return false;">'.html_print_image('images/copy.png', true).'</a>';
+			onClick="if (!confirm(\''.__('Are you sure?').'\')) return false;">'.html_print_image('images/copy.png', true, ['class' => 'invert_filter']).'</a>';
         $data['action'] .= '<a href="index.php?sec='.$sec.'&sec2=godmode/alerts/alert_commands&delete_command=1&id='.$command['id'].'&pure='.$pure.'"
-			onClick="if (!confirm(\''.__('Are you sure?').'\')) return false;">'.html_print_image('images/cross.png', true).'</a>';
+			onClick="if (!confirm(\''.__('Are you sure?').'\')) return false;">'.html_print_image('images/cross.png', true, ['class' => 'invert_filter']).'</a>';
         $data['action'] .= '</span>';
     }
 
     array_push($table->data, $data);
 }
 
-if (count($table->data) > 0) {
+ui_pagination($total_commands, $url);
+if (isset($data) === true && count($table->data) > 0) {
     html_print_table($table);
+    ui_pagination($total_commands, $url, 0, 0, false, 'offset', true, 'pagination-bottom');
 } else {
-    ui_print_info_message(['no_close' => true, 'message' => __('No alert commands configured') ]);
+    ui_print_info_message(
+        [
+            'no_close' => true,
+            'message'  => __('No alert commands configured'),
+        ]
+    );
 }
 
-if ($is_central_policies_on_node === false) {
+if ($is_central_policies_on_node === false && check_acl_restricted_all($config['id_user'], $command['id_group'], 'PM')) {
     echo '<div class="action-buttons" style="width: '.$table->width.'">';
     echo '<form method="post" action="index.php?sec='.$sec.'&sec2=godmode/alerts/configure_alert_command&pure='.$pure.'">';
     html_print_submit_button(__('Create'), 'create', false, 'class="sub next"');
@@ -608,3 +709,26 @@ if ($is_central_policies_on_node === false) {
 }
 
 enterprise_hook('close_meta_frame');
+
+?>
+
+<script type="text/javascript">
+    $(document).ready(function () {
+        dialog_message("#message_permissions");
+    });
+
+    function dialog_message(message) {
+      $(message)
+        .css("display", "inline")
+        .dialog({
+          modal: true,
+          width: "400px",
+          buttons: {
+            Close: function() {
+              $(this).dialog("close");
+            }
+          }
+        });
+    }
+
+</script>
